Remove plotting from MMS
